+++ src/Plugins/AppleGameCenterPlugin/AppleGameCenterNative.h
@@ -12,6 +12,6 @@
 - (BOOL) loadCompletedAchievements:(void(^ _Nonnull)(NSError * _Nullable, NSArray* _Nullable))handler;
 
 - (BOOL) reportScore:(NSString * _Nonnull)identifier score:(int64_t)score response:(void(^ _Nonnull)(NSError * _Nullable))handler;
-- (BOOL) reportAchievementIdentifier:(NSString * _Nonnull)identifier percentComplete:(float)percent withBanner:(BOOL)banner response:(void(^ _Nonnull)(NSError * _Nullable))handler;
+- (BOOL) reportAchievementIdentifier:(NSString * _Nonnull)identifier percentComplete:(double)percent withBanner:(BOOL)banner response:(void(^ _Nonnull)(NSError * _Nullable))handler;
 
 @end
